+++ packages/sit-onyx/src/components/OnyxListbox/OnyxListbox.vue
@@ -43,6 +43,14 @@ const slots = defineSlots<{
    * to e.g. show a button to load more options instead of lazy loading on scroll.
    */
   optionsEnd?(): unknown;
+  /**
+   * Optional header content to display above the options.
+   */
+  header?(): unknown;
+  /**
+   * Optional footer content to display below the options (will replace `message` property).
+   */
+  footer?(): unknown;
 }>();
 
 const { t } = injectI18n();
@@ -166,7 +174,16 @@ watchEffect(() => {
 </script>
 
 <template>
-  <div class="onyx-listbox" :aria-busy="props.loading">
+  <div
+    class="onyx-listbox"
+    :class="{
+      'onyx-listbox--with-header': !!slots.header,
+      'onyx-listbox--with-footer': !!slots.footer,
+    }"
+    :aria-busy="props.loading"
+  >
+    <slot name="header"></slot>
+
     <div v-if="props.loading" class="onyx-listbox__slot onyx-listbox__slot--loading">
       <OnyxLoadingIndicator class="onyx-listbox__loading" />
     </div>
@@ -224,6 +241,8 @@ watchEffect(() => {
           "
           :multiple="props.multiple"
           :active="option.value === activeOption"
+          :icon="option.icon"
+          :color="option.color"
         >
           {{ option.label }}
         </OnyxListboxOption>
@@ -237,9 +256,12 @@ watchEffect(() => {
         <slot name="optionsEnd"></slot>
       </div>
     </div>
-    <span v-if="props.message" class="onyx-listbox__message onyx-text--small">
-      {{ props.message }}
-    </span>
+
+    <slot name="footer">
+      <span v-if="props.message" class="onyx-listbox__message onyx-text--small">
+        {{ props.message }}
+      </span>
+    </slot>
   </div>
 </template>
 
@@ -261,6 +283,14 @@ watchEffect(() => {
     min-width: var(--onyx-spacing-4xl);
     max-width: 20rem;
     font-family: var(--onyx-font-family);
+
+    &--with-header {
+      padding-top: 0;
+    }
+
+    &--with-footer {
+      padding-bottom: 0;
+    }
 
     &__wrapper {
       max-height: calc(var(--max-options) * var(--option-height));
@@ -301,10 +331,6 @@ watchEffect(() => {
       height: var(--option-height);
     }
 
-    .onyx-listbox-option {
-      height: var(--option-height);
-    }
-
     &:has(&__wrapper:focus-visible) {
       outline: 0.25rem solid var(--onyx-color-base-primary-200);
     }
@@ -323,6 +349,10 @@ watchEffect(() => {
 
     &__loading {
       color: var(--onyx-color-text-icons-primary-intense);
+    }
+
+    .onyx-empty {
+      max-width: 100%;
     }
   }
 }