+++ packages/sit-onyx/src/components/OnyxListboxOption/OnyxListboxOption.vue
@@ -1,7 +1,12 @@
 <script lang="ts" setup>
+import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import type { OnyxListboxOptionProps } from "./types";
 
-const props = defineProps<OnyxListboxOptionProps>();
+const props = withDefaults(defineProps<OnyxListboxOptionProps>(), {
+  active: false,
+  multiple: false,
+  color: "primary",
+});
 
 defineSlots<{
   /**
@@ -12,7 +17,13 @@ defineSlots<{
 </script>
 
 <template>
-  <li class="onyx-listbox-option" :class="{ 'onyx-listbox-option--active': props.active }">
+  <li
+    class="onyx-listbox-option"
+    :class="{
+      'onyx-listbox-option--active': props.active,
+      'onyx-listbox-option--danger': props.color === 'danger',
+    }"
+  >
     <input
       v-if="props.multiple"
       :checked="!!$attrs['aria-checked']"
@@ -24,6 +35,8 @@ defineSlots<{
       class="onyx-listbox-option__checkbox"
       type="checkbox"
     />
+
+    <OnyxIcon v-if="props.icon" :icon="props.icon" />
 
     <span class="onyx-truncation-ellipsis">
       <slot></slot>
@@ -39,8 +52,13 @@ defineSlots<{
   @include checkbox.variables();
 
   @include layers.component() {
+    --onyx-listbox-option-color: var(--onyx-color-text-icons-neutral-intense);
+    --onyx-listbox-option-color-selected: var(--onyx-color-text-icons-primary-bold);
+    --onyx-listbox-option-background-hover: var(--onyx-color-base-primary-100);
+    --onyx-listbox-option-background-selected: var(--onyx-color-base-primary-200);
+
     font-family: var(--onyx-font-family);
-    color: var(--onyx-color-text-icons-neutral-intense);
+    color: var(--onyx-listbox-option-color);
     padding: var(--onyx-spacing-2xs) var(--onyx-spacing-sm);
     background-color: var(--onyx-color-base-background-blank);
     font-weight: 400;
@@ -66,31 +84,40 @@ defineSlots<{
 
       &:hover,
       &.onyx-listbox-option--active {
-        background-color: var(--onyx-color-base-primary-100);
+        background-color: var(--onyx-listbox-option-background-hover);
       }
 
       // single select
       &[aria-selected="true"] {
-        background-color: var(--onyx-color-base-primary-200);
+        background-color: var(--onyx-listbox-option-background-selected);
 
         &:hover,
         &.onyx-listbox-option--active {
-          color: var(--onyx-color-text-icons-primary-bold);
+          color: var(--onyx-listbox-option-color-selected);
         }
       }
 
       // single + multiselect
       &[aria-checked="true"] {
+        background-color: var(--onyx-listbox-option-background-selected);
+
         &:hover,
         &.onyx-listbox-option--active {
-          background-color: var(--onyx-color-base-primary-100);
-          color: var(--onyx-color-text-icons-primary-bold);
+          background-color: var(--onyx-listbox-option-background-selected);
+          color: var(--onyx-listbox-option-color-selected);
         }
       }
     }
 
     &[aria-disabled="true"] {
       color: var(--onyx-color-text-icons-neutral-soft);
+    }
+
+    &--danger {
+      --onyx-listbox-option-color: var(--onyx-color-text-icons-danger-intense);
+      --onyx-listbox-option-color-selected: var(--onyx-color-text-icons-danger-intense);
+      --onyx-listbox-option-background-hover: var(--onyx-color-base-danger-100);
+      --onyx-listbox-option-background-selected: var(--onyx-color-base-danger-200);
     }
   }
 }