+++ packages/sit-onyx/src/components/OnyxNavItem/OnyxNavItem.vue
@@ -113,6 +113,7 @@ const shouldShowExternalIcon = computed(() => {
 
     &--active {
       color: var(--onyx-color-text-icons-secondary-intense);
+      font-weight: 600;
 
       &::after {
         content: " ";