+++ packages/sit-onyx/src/index.ts
@@ -40,6 +40,9 @@ export * from "./components/OnyxLink/types";
 export { default as OnyxListbox } from "./components/OnyxListbox/OnyxListbox.vue";
 export * from "./components/OnyxListbox/types";
 
+export { default as OnyxListboxOption } from "./components/OnyxListboxOption/OnyxListboxOption.vue";
+export * from "./components/OnyxListboxOption/types";
+
 export { default as OnyxLoadingIndicator } from "./components/OnyxLoadingIndicator/OnyxLoadingIndicator.vue";
 export * from "./components/OnyxLoadingIndicator/types";
 
@@ -71,6 +74,9 @@ export * from "./components/OnyxTooltip/types";
 
 export { default as OnyxTag } from "./components/OnyxTag/OnyxTag.vue";
 export * from "./components/OnyxTag/types";
+
+export { default as OnyxUserMenu } from "./components/OnyxUserMenu/OnyxUserMenu.vue";
+export * from "./components/OnyxUserMenu/types";
 
 export * from "./composables/density";
 export * from "./composables/scrollEnd";