+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # alpha-test-app
 
+## 0.1.0-alpha.89
+
+### Patch Changes
+
+- Updated dependencies [2b48da9]
+- Updated dependencies [2b48da9]
+  - sit-onyx@1.0.0-alpha.92
+
+## 0.1.0-alpha.88
+
+### Patch Changes
+
+- Updated dependencies [93ad5a9]
+- Updated dependencies [1eb0528]
+  - sit-onyx@1.0.0-alpha.91
+
 ## 0.1.0-alpha.87
 
 ### Patch Changes