+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,21 @@
 # playground
 
+## 0.0.1-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [2b48da9]
+- Updated dependencies [2b48da9]
+  - sit-onyx@1.0.0-alpha.92
+
+## 0.0.1-alpha.30
+
+### Patch Changes
+
+- Updated dependencies [93ad5a9]
+- Updated dependencies [1eb0528]
+  - sit-onyx@1.0.0-alpha.91
+
 ## 0.0.1-alpha.29
 
 ### Patch Changes