+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,19 @@
 # sit-onyx
 
+## 1.0.0-alpha.92
+
+### Minor Changes
+
+- 2b48da9: feat: add basic `OnyxUserMenu` component
+- 2b48da9: feat(OnyxListbox): support icon and danger color for options
+
+## 1.0.0-alpha.91
+
+### Patch Changes
+
+- 93ad5a9: Fix hover colors on OnyxListbox
+- 1eb0528: Fix OnyxNavItem font-weight for active state
+
 ## 1.0.0-alpha.90
 
 ### Minor Changes